% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyRDT.R
\name{shinyRDT-shiny}
\alias{shinyRDT-shiny}
\alias{shinyRDTOutput}
\alias{renderShinyRDT}
\alias{shinyRDT_html}
\title{Shiny bindings for shinyRDT}
\usage{
shinyRDTOutput(outputId, width = "100\%", height = "400px")

renderShinyRDT(expr, env = parent.frame(), quoted = FALSE)

shinyRDT_html(id, style, class, ...)
}
\arguments{
\item{outputId}{output variable to read from}

\item{expr}{An expression that generates a shinyRDT}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using shinyRDT within Shiny
applications and interactive Rmd documents.
}
